requirements.txt
asgiref==3.2.10
Django==2.2.5
django-cors-headers==3.4.0
django-filter==2.2.0
django-health-check==3.11.0
djangorestframework==3.10.3
djangorestframework-jwt==1.11.0
djangorestframework-simplejwt==4.3.0
gunicorn==19.9.0
Markdown==3.1.1
PyJWT==1.7.1
pytz==2019.2
sqlparse==0.3.0
mysqlclient
python-resize-image
pycoingecko
